edge_index needs long?

+++ src/dinobots/data_structures/PolarGraph.py
@@ -171,7 +171,7 @@ class PolarGraph:
         return Data(
             x=torch.tensor(x, dtype=torch.float32) if x is not None else None,
             edge_index=torch.transpose(
-                torch.tensor(edge_index, dtype=torch.float32), 0, 1
+                torch.tensor(edge_index, dtype=torch.long), 0, 1
             )
             if edge_index is not None
             else None,
